Delegate cob spec file service to (trivial) FileController
