+++ src/main/java/HttpServer/controller/Controller.java
@@ -11,6 +11,10 @@ public class Controller implements IController {
         this.router = new Router(new Routes());
     }
 
+    public Controller(Router router) {
+        this.router = router;
+    }
+
     @Override
     public Router getRouter() {
         return router;
